Put writer mode functions in separate file
